backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.2",
    "openai>=1.50.0",
    "pydantic==2.6.1",
    "python-dotenv==1.0.1",
    "uvicorn==0.27.1",
    "sqlalchemy>=2.0.0",
]
